repl: fix locals and globals issues
Previously variables defined in the REPL were not accessible to functions or
methods, because rlpython saved them into locals().

    >>> a = 10
    >>> def foo():
    ...     print(a)
    ...
    >>> a
    10
    >>> foo()
    Traceback (most recent call last):
      File "/home/fls/opt/python3.9/lib/python3.9/site-packages/rlpython/runtimes/python_runtime.py", line 211, in run
        return_value = run_code(code, self.globals, self.locals)
      File "<input>", line 1, in <module>
      File "<input>", line 2, in foo
    NameError: name 'a' is not defined
    >>>

The default Python REPL saves variables into globals() which avoids this kind
of problems altogether.

This patch merges rlpythons locals() and globals() to fix this issue and behave
more like the standard Python REPL.

Signed-off-by: Florian Scherf <mail@florianscherf.de>