repl: make templating configurable
rlpythons templating feature makes using Jinja2 templates impossible, because
if you try to type in Jinja2 syntax in a string, rlpython tries to evaluate
it before declaring the string. This results in confusing NameErrors.

This Patch adds an switch, if templating should be applied to the command line,
as REPL variable.

Signed-off-by: Florian Scherf <mail@florianscherf.de>